This fixes the build on the current nightly, and also #157, by simply removing all calls to `Thread::scoped` from `src/ast.rs`. All tests still pass, and no regressions have been found after some initial testing.