@tafia 2015-02-20
For my information #157 raised some performance issue. Did you have tile to
make some benchmark to see if it changes something ?
On 20 Feb 2015 02:09, "Renato Zannon" <notifications@github.com> wrote:

> This fixes the build on the current nightly, and also #157
> <https://github.com/phildawes/racer/issues/157>, by simply removing all
> calls to Thread::scoped from src/ast.rs. All tests still pass, and no
> regressions have been found after some initial testing.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/phildawes/racer/pull/159
> Commit Summary
>
>    - Remove uses of Thread::scoped from src/ast.rs - fixes #157
>    - Fix feature flags (kills most warnings)
>
> File Changes
>
>    - *M* src/main.rs
>    <https://github.com/phildawes/racer/pull/159/files#diff-0> (2)
>    - *M* src/racer/ast.rs
>    <https://github.com/phildawes/racer/pull/159/files#diff-1> (230)
>
> Patch Links:
>
>    - https://github.com/phildawes/racer/pull/159.patch
>    - https://github.com/phildawes/racer/pull/159.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/phildawes/racer/pull/159>.
>