@renato-zannon 2015-02-20
Just did!

```
$ cat bench.sh
N=100
for i in $(seq 1 $N); do
  target/release/racer complete std::old_io:: >/dev/null
done

# Without the PR:
$ rustc --version
rustc 1.0.0-nightly (b63cee4a1 2015-02-14 17:01:11 +0000)

$ time ./bench.sh
real    0m18.654s
user    0m10.353s
sys     0m8.313s

$ cargo bench
# <snip>
test racer::bench::bench_code_chunks     ... bench:    200484 ns/iter (+/- 11952)
test racer::bench::bench_iter_stmts      ... bench:    478272 ns/iter (+/- 23615)
test racer::bench::bench_mask_comments   ... bench:    296777 ns/iter (+/- 11999)
test racer::bench::bench_mask_sub_scopes ... bench:    399283 ns/iter (+/- 15699)

# With the PR:

$ rustc --version
rustc 1.0.0-nightly (dfc5c0f1e 2015-02-18) (built 2015-02-19)

$ time ./bench.sh
real    0m11.323s
user    0m7.189s
sys     0m4.117s

$ cargo bench
# <snip>
test racer::bench::bench_code_chunks     ... bench:    198126 ns/iter (+/- 6881)
test racer::bench::bench_iter_stmts      ... bench:    483020 ns/iter (+/- 20938)
test racer::bench::bench_mask_comments   ... bench:    303417 ns/iter (+/- 11246)
test racer::bench::bench_mask_sub_scopes ... bench:    388178 ns/iter (+/- 14858)
```

So it seems that it didn't influence the benchmarks too much (maybe they don't hit this code path?), but there's a big difference on runtime.

That said, the main reason for this PR is that the behavior racer was counting on - using scoped threads to catch panics - isn't supported by rust anymore, and the code needed to be changed to reflect that anyway. The performance improvement is just a nice side effect :)