@phildawes 2015-02-20
Quick (relatively unscientific) test on a linux laptop:
```rust
let v = Vec::new();
v.<complete>
```

Before Renato's patch:
```
% time ./target/release/racer complete 2 2 /tmp/a.rs
real	0m0.635s
user	0m0.331s
sys	0m0.309s
```

After:
```
% time ./target/release/racer complete 2 2 /tmp/a.rs
real	0m0.323s
user	0m0.187s
sys	0m0.135s
```
Nice!